After Disconnecting Ethernet auto reconnect  is not Working in brige support.
:Release Notes:
After Disconnecting Ethernet auto reconnect  is not Working in brige support.

:Detailed Notes:
Informing the user about the type of network supports we are providing i.e NAT, Bridge support.
Intimating the user to keep a static internet connection while bridge support is opted, as unstable connection may cause random behavior.

:Testing Performed:
Connection via ssh.
ssh connection check when the emulator is launhced with the bridge support.
ssh connection check when the emulator is launhced with the NAT support.
Internet connection check when the emulator is launhced with the NAT support.
Internet connection check when the emulator is launhced with the Bridge support.

:QA Notes:

:Issues Addressed:
[PLAT-71531] [QEMU Emulator] : After Disconnecting wired Connection (Ethernet) is not Working .